requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.110.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
orjson==3.11.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.9
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
uvicorn==0.29.0
watchfiles==1.1.0
websockets==15.0.1
mysql-connector-python==8.0.33
